Refactored out data type modules
Refactored out the data type modules but left REPL alone for now.
