+++ .gitignore
@@ -33,4 +33,4 @@
 
 # Miscellaneous
 *.swp
-ptscheme.
+ptscheme
