+++ Makefile
@@ -8,21 +8,32 @@
 
 vpath %.c src
 vpath %.h include
+
 CPPFLAGS = -std=c99 -Wall -ggdb -I include
+objects  = ptscheme.o memmanager.o lispbool.o lispchar.o lispint.o lispstr.o lisperr.o
 
 # Standard targets
 
-all: ptscheme clean
+all: ptscheme
 
 # Define all our non-standard targets and their dependencies
 
-ptscheme: ptscheme.o
+ptscheme: $(objects)
+	cc $(CPPFLAGS) -o ptscheme $(objects)
+
+ptscheme.o: ptscheme.h memmanager.h
+memmanager.o: ptscheme.h memmanager.h
+lispbool.o: ptscheme.h memmanager.h lispbool.h
+lispchar.o: ptscheme.h memmanager.h lispchar.h
+lispint.o: ptscheme.h memmanager.h lispint.h
+lispstr.o: ptscheme.h memmanager.h lispstr.h
+lisperr.o: ptscheme.h memmanager.h lisperr.h
 
 # clean - Implement standard 'clean' target to clean up after ourselves
 
 .PHONY: clean
 clean:
-	rm -f *.o
+	rm -f ptscheme $(objects)
 
 # help - Provide some basic help capabilities by listing available targets
