+++ include/lispstr.h
@@ -1,0 +1,9 @@
+/* lispstr.h */
+
+#ifndef FILE_LISPSTR_SEEN
+#define FILE_LISPSTR_SEEN
+
+object *make_string(char *value);
+bool is_string(object *obj);
+
+#endif // !FILE_LISPSTR_SEEN
