+++ include/memmanager.h
@@ -1,0 +1,10 @@
+/* memmanager.h */
+
+#ifndef FILE_MEMMANAGER_SEEN
+#define FILE_MEMMANAGER_SEEN
+
+/* Lisp object constructors */
+
+extern object *alloc_object(void);
+
+#endif // !FILE_MEMMANAGER_SEEN
