+++ src/ptscheme.c
@@ -31,150 +31,13 @@
 #include <stdbool.h>
 #include <string.h>
 #include <ctype.h>
-
-/* Const declarations */
-
-const int BUFFER_MAX = 1000;
-
-/* Type declarations */
-
-typedef enum {
-	BOOLEAN,
-	CHARACTER,
-	FIXNUM,
-	STRING,
-	ERROR
-} object_type;
-
-typedef struct object {
-	object_type type;
-	union {
-		struct {
-			char value;
-		} boolean;
-		struct {
-			char value;
-		} character;
-		struct {
-			long value;
-		} fixnum;
-		struct {
-			char *value;
-		} string;
-		struct {
-			long error_num;
-			char *error_msg;
-		} error;
-	} data;
-} object;
-
-/* Lisp object constructors */
-
-object *alloc_object(void) {
-	object *obj;
-	obj = malloc(sizeof(object));
-	if (obj == NULL) {
-		fprintf(stderr, "Out of memory\n");
-		exit(EXIT_FAILURE);
-	}
-	return obj;
-}
-
-// Boolean native type
-
-object *make_boolean(char value) {
-
-	// The boolean objects are singletons
-
-	static object *bool_true;
-	static object *bool_false;
-
-	if (bool_true == NULL) {
-		bool_true = alloc_object();
-		bool_true->type = BOOLEAN;
-		bool_true->data.boolean.value = true;
-	}
-	if (bool_false == NULL) {
-		bool_false = alloc_object();
-		bool_false->type = BOOLEAN;
-		bool_false->data.boolean.value = false;
-	}
-	return ((value) ? bool_true : bool_false);
-}
-
-bool is_boolean(object *obj) {
-	return (obj->type == BOOLEAN);
-}
-
-bool is_false(object *obj) {
-	return (obj->data.boolean.value == false);
-}
-
-bool is_true(object *obj) {
-	return (obj->data.boolean.value == true);
-}
-
-// Character native type
-
-object *make_character(char value) {
-	object *obj;
-	obj = alloc_object();
-	obj->type = CHARACTER;
-	obj->data.character.value = value;
-	return obj;
-}
-
-bool is_character(object *obj) {
-	return (obj->type == CHARACTER);
-}
-
-// Integer native type
-
-object *make_fixnum(long value) {
-	object *obj;
-	obj = alloc_object();
-	obj->type = FIXNUM;
-	obj->data.fixnum.value = value;
-	return obj;
-}
-
-bool is_fixnum(object *obj) {
-	return (obj->type == FIXNUM);
-}
-
-// String native type
-
-object *make_string(char *value) {
-	object *obj;
-	obj = alloc_object();
-	obj->type = STRING;
-	obj->data.string.value = malloc(strlen(value) + 1);
-	if (obj->data.string.value == NULL) {
-		fprintf(stderr, "Out of memory\n");
-		exit(EXIT_FAILURE);
-	}
-	strcpy(obj->data.string.value, value);
-	return obj;
-}
-
-bool is_string(object *obj) {
-	return (obj->type == STRING);
-}
-
-// Error native type
-
-object *make_error(long error_num, char *error_msg) {
-	object *obj;
-	obj = alloc_object();
-	obj->type = ERROR;
-	obj->data.error.error_num = error_num;
-	obj->data.error.error_msg = error_msg;
-	return obj;
-}
-
-bool is_error(object *obj) {
-	return (obj->type == ERROR);
-}
+#include "ptscheme.h"
+#include "memmanager.h"
+#include "lispbool.h"
+#include "lispchar.h"
+#include "lispint.h"
+#include "lispstr.h"
+#include "lisperr.h"
 
 /* REPL - Read */
